requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
supabase==2.10.0
pydantic==2.10.3
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.24.0
hypothesis==6.122.3
